requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
appdirs==1.4.4
    # via
    #   -r requirements/base.txt
    #   fs
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
attrs==24.3.0
    # via -r requirements/base.txt
bleach==6.2.0
    # via -r requirements/base.txt
boto3==1.35.82
    # via
    #   -r requirements/base.txt
    #   fs-s3fs
botocore==1.35.82
    # via
    #   -r requirements/base.txt
    #   boto3
    #   s3transfer
certifi==2024.12.14
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   pynacl
charset-normalizer==3.4.0
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.7
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django==4.2.17
    # via
    #   -r requirements/base.txt
    #   django-appconf
    #   django-config-models
    #   django-crum
    #   django-filter
    #   django-statici18n
    #   django-waffle
    #   djangorestframework
    #   edx-django-utils
    #   edx-i18n-tools
    #   jsonfield
    #   openedx-django-pyfs
    #   openedx-filters
django-appconf==1.0.6
    # via
    #   -r requirements/base.txt
    #   django-statici18n
django-config-models==2.7.0
    # via -r requirements/base.txt
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-filter==24.3
    # via -r requirements/base.txt
django-statici18n==2.6.0
    # via -r requirements/base.txt
django-waffle==4.2.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
djangorestframework==3.15.2
    # via
    #   -r requirements/base.txt
    #   django-config-models
dnspython==2.7.0
    # via
    #   -r requirements/base.txt
    #   pymongo
edx-ccx-keys==2.0.2
    # via -r requirements/base.txt
edx-django-utils==7.1.0
    # via
    #   -r requirements/base.txt
    #   django-config-models
edx-i18n-tools==1.6.3
    # via -r requirements/dev.in
edx-opaque-keys[django]==2.11.0
    # via
    #   -r requirements/base.txt
    #   edx-ccx-keys
fs==2.4.16
    # via
    #   -r requirements/base.txt
    #   fs-s3fs
    #   openedx-django-pyfs
    #   xblock
fs-s3fs==1.1.1
    # via
    #   -r requirements/base.txt
    #   openedx-django-pyfs
future==1.0.0
    # via
    #   -r requirements/base.txt
    #   pyjwkest
idna==3.10
    # via
    #   -r requirements/base.txt
    #   requests
jmespath==1.0.1
    # via
    #   -r requirements/base.txt
    #   boto3
    #   botocore
jsonfield==3.1.0
    # via -r requirements/base.txt
lazy==1.6
    # via -r requirements/base.txt
lxml[html-clean]==5.3.0
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
    #   lxml-html-clean
    #   xblock
lxml-html-clean==0.4.1
    # via lxml
mako==1.3.8
    # via
    #   -r requirements/base.txt
    #   xblock
markupsafe==3.0.2
    # via
    #   -r requirements/base.txt
    #   mako
    #   xblock
newrelic==10.4.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
oauthlib==3.2.2
    # via -r requirements/base.txt
openedx-django-pyfs==3.7.0
    # via -r requirements/base.txt
openedx-filters==1.12.0
    # via -r requirements/base.txt
path==16.16.0
    # via edx-i18n-tools
pbr==6.1.0
    # via
    #   -r requirements/base.txt
    #   stevedore
polib==1.2.0
    # via edx-i18n-tools
psutil==6.1.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pycryptodomex==3.21.0
    # via
    #   -r requirements/base.txt
    #   pyjwkest
pyjwkest==1.4.2
    # via -r requirements/base.txt
pyjwt==2.6.0
    # via -r requirements/base.txt
pymongo==4.10.1
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.txt
    #   botocore
    #   xblock
pytz==2024.2
    # via
    #   -r requirements/base.txt
    #   xblock
pyyaml==6.0.2
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
    #   xblock
requests==2.32.3
    # via
    #   -r requirements/base.txt
    #   pyjwkest
s3transfer==0.10.4
    # via
    #   -r requirements/base.txt
    #   boto3
simplejson==3.19.3
    # via
    #   -r requirements/base.txt
    #   xblock
six==1.17.0
    # via
    #   -r requirements/base.txt
    #   edx-ccx-keys
    #   fs
    #   fs-s3fs
    #   pyjwkest
    #   python-dateutil
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-opaque-keys
typing-extensions==4.12.2
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
urllib3==1.26.20
    # via
    #   -r requirements/base.txt
    #   botocore
    #   requests
web-fragments==2.2.0
    # via
    #   -r requirements/base.txt
    #   xblock
webencodings==0.5.1
    # via
    #   -r requirements/base.txt
    #   bleach
webob==1.8.9
    # via
    #   -r requirements/base.txt
    #   xblock
xblock==5.1.0
    # via -r requirements/base.txt

# The following packages are considered to be unsafe in a requirements file:
# setuptools
